fixup! fixup! [cluster-test] Create a test for compatibility between commits
